MCR-2497 Outsourced api check and detail filter to extra filter
